% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitVSP.R
\name{fitVSP}
\alias{fitVSP}
\title{Fit V-Spline model}
\usage{
fitVSP(dat, pa = c(1, 1))
}
\description{
Fit V-Spline model
}
\references{
Z. Cao, D. Bryant, C. Fox, T. Molten and M. Parry. "V-Spline: an Adaptive Smoothing Spline
for Trajectory Reconstruction" arXiv preprint arXiv:1803.07184 (2018).
}
