% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheta.R
\name{getTheta}
\alias{getTheta}
\title{The estimated parameter \eqn{\theta}}
\usage{
getTheta(data_frame, pa, matrix_list)
}
\arguments{
\item{data_frame}{the data frame}

\item{pa}{the parameters \eqn{\lambda} and \eqn{\gamma}, which were found by \code{loocvVSP()} function.}

\item{matrix_list}{the initialized matrices found by \code{iniVSPmatrices()}.}
}
\description{
to get the estimated parameter \eqn{\hat{\theta}}.
\deqn{f(x; \Phi) = \sum_{j = 1}^g \pi_j f_j(x; \theta_j)}
}
\references{
Z. Cao, D. Bryant, C. Fox, T. Molten and M. Parry. "V-Spline: an Adaptive Smoothing Spline
for Trajectory Reconstruction" arXiv preprint arXiv:1803.07184 (2018).
}
\keyword{theta}
